backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
python-jose
passlib[bcrypt]
psycopg2-binary
python-multipart
jinja2
requests
python-dotenv
pydantic[email]
langgraph
groq
python-dotenv
pytest
kafka-python
websockets 
textblob
twilio
faster-whisper
loguru